\name{AICc}
\alias{AICc}
\title{Compute AICc}
\usage{
AICc(mod, ...)
}
\arguments{
  \item{mod}{\code{midas_r} model}

  \item{...}{additional parameters}
}
\value{
a computed AICc value, a number.
}
\description{
Function to compute AICc information criteria for a given
model
}
\details{
A generic function for calculating AICc. It is implemented
for \link{midas_r} package. For more examples see package
AICcmodavg.
}
\author{
Vaidotas Zemlys
}

