\name{almonp.gradient}
\alias{almonp.gradient}
\title{Gradient function for Almon polynomial MIDAS weights}
\usage{
  almonp.gradient(p, d)
}
\arguments{
  \item{p}{vector of parameters for Almon polynomial
  specification}

  \item{d}{number of coefficients}
}
\value{
  vector of coefficients
}
\description{
  Calculate gradient for Almon polynomial MIDAS weights
  specification
}
\author{
  Vaidotas Zemlys
}

