% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{almonp}
\alias{almonp}
\title{Almon polynomial MIDAS weights specification}
\usage{
almonp(p, d, m)
}
\arguments{
\item{p}{parameters for Almon polynomial weights}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Calculate Almon polynomial MIDAS weights
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}

