% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{almonp_gradient}
\alias{almonp_gradient}
\title{Gradient function for Almon polynomial MIDAS weights}
\usage{
almonp_gradient(p, d, m)
}
\arguments{
\item{p}{vector of parameters for Almon polynomial specification}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Calculate gradient for Almon polynomial MIDAS weights specification
}
\author{
Vaidotas Zemlys
}
