% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_nlpr_methods.R
\name{deviance.midas_nlpr}
\alias{deviance.midas_nlpr}
\title{Non-linear parametric MIDAS regression model deviance}
\usage{
\method{deviance}{midas_nlpr}(object, ...)
}
\arguments{
\item{object}{a \code{\link{midas_r}} object}

\item{...}{currently nothing}
}
\value{
The sum of squared residuals
}
\description{
Returns the deviance of a fitted MIDAS regression object
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
