% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midaslag.R
\name{dmls}
\alias{dmls}
\title{MIDAS lag structure for unit root processes}
\usage{
dmls(x, k, m, ...)
}
\arguments{
\item{x}{a vector}

\item{k}{maximal lag order}

\item{m}{frequency ratio}

\item{...}{further arguments used in fitting MIDAS regression}
}
\value{
a matrix containing the first differences and the lag k+1.
}
\description{
Prepares MIDAS lag structure for unit root processes
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}

