\name{embedlf_coef}
\alias{embedlf_coef}
\title{Return the coefficients for embedlf variables}
\usage{
  embedlf_coef(x)
}
\arguments{
  \item{x}{an output from \code{\link{midas_u}}}
}
\value{
  a vector
}
\description{
  Extracts the coefficients and returns those coefficients
  which name has string \code{embedlf} in it.
}
\author{
  Vaidotas Zemlys
}

