% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_r_methods.R
\name{extract.midas_r}
\alias{extract.midas_r}
\title{Extract coefficients and GOF measures from MIDAS regression object}
\usage{
\method{extract}{midas_r}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE, include.rmse = TRUE, ...)
}
\arguments{
\item{model}{a MIDAS regression object}

\item{include.rsquared, }{If available: should R-squared be reported?}

\item{include.adjrs}{If available: should the adjusted R-squared be reported?}

\item{include.nobs}{If available: should the number of observations be reported?}

\item{include.rmse}{If available: should the root-mean-square error (= residual standard deviation) be reported?}

\item{...}{additional parameters passed to summary}
}
\value{
texreg object
}
\description{
Extract coefficients and GOF measures from MIDAS regression object
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
