% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_nlpr_methods.R
\name{fitted.midas_nlpr}
\alias{fitted.midas_nlpr}
\title{Fitted values for non-linear parametric MIDAS regression model}
\usage{
\method{fitted}{midas_nlpr}(object, ...)
}
\arguments{
\item{object}{a \code{\link{midas_r}} object}

\item{...}{currently nothing}
}
\value{
the vector of fitted values
}
\description{
Returns the fitted values of a fitted non-linear parametric MIDAS regression object
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
