% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_sp_methods.R
\name{fitted.midas_sp}
\alias{fitted.midas_sp}
\title{Fitted values for semi-parametric MIDAS regression model}
\usage{
\method{fitted}{midas_sp}(object, ...)
}
\arguments{
\item{object}{a \code{\link{midas_r}} object}

\item{...}{currently nothing}
}
\value{
the vector of fitted values
}
\description{
Returns the fitted values of a fitted semi-parametric MIDAS regression object
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
