% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midaslag.R
\name{fmls}
\alias{fmls}
\title{Full MIDAS lag structure}
\usage{
fmls(x, k, m, ...)
}
\arguments{
\item{x}{a vector}

\item{k}{maximum lag order}

\item{m}{frequency ratio}

\item{...}{further arguments}
}
\value{
a matrix containing the lags
}
\description{
Create a matrix of MIDAS lags, including contemporaneous lag up to selected order.
}
\details{
This is a convenience function, it calls \code{link{mls}} to perform actual calculations.
}
\seealso{
mls
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
