\name{forecast.midas_r}
\alias{forecast.midas_r}
\title{Forecast MIDAS regression}
\usage{
  \method{forecast}{midas_r} (object, newdata = NULL, ...)
}
\arguments{
  \item{object}{midas_r object}

  \item{newdata}{newdata}

  \item{...}{additional arguments, not used}
}
\value{
  a vector of forecasts
}
\description{
  Forecasts MIDAS regression. Differs from \code{predict},
  in that it respects history
}
\details{
  Add later
}
\author{
  Vaidotas Zemlys
}

