\name{forecast}
\alias{forecast}
\title{Forecasti midas_r object}
\usage{
  forecast(object, ...)
}
\arguments{
  \item{object}{object to forecast}

  \item{...}{additional arguments}
}
\value{
  a forecast
}
\description{
  A generic function for forecasting R objects.
}
\details{
  Added for compatibility with \code{forecast} package
}

