% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{genexp}
\alias{genexp}
\title{Generalized exponential MIDAS coefficients}
\usage{
genexp(p, d, m)
}
\arguments{
\item{p}{a vector of parameters}

\item{d}{number of coefficients}

\item{m}{the frequency, currently ignored}
}
\value{
a vector of coefficients
}
\description{
Calculates the MIDAS coefficients for generalized exponential MIDAS lag specification
}
\details{
Generalized exponential MIDAS lag specification is a generalization of exponential
Almon lag. It is defined as a product of first order polynomial with exponent
of the second order polynomial. This spefication was used by V. Kvedaras and
V. Zemlys (2012).
}
\references{
Kvedaras V., Zemlys, V. \emph{Testing the functional constraints on
parameters in regressions with variables of different frequency} Economics Letters 116 (2012) 250-254
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
