% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lagspec.R
\name{genexp_gradient}
\alias{genexp_gradient}
\title{Gradient of feneralized exponential MIDAS coefficient generating function}
\usage{
genexp_gradient(p, d, m)
}
\arguments{
\item{p}{a vector of parameters}

\item{d}{number of coefficients}

\item{m}{the frequency, currently ignored}
}
\value{
a vector of coefficients
}
\description{
Calculates the gradient of generalized exponential MIDAS lag specification
}
\details{
Generalized exponential MIDAS lag specification is a generalization of exponential
Almon lag. It is defined as a product of first order polynomial with exponent
of the second order polynomial. This spefication was used by V. Kvedaras and
V. Zemlys (2012).
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
\references{
Kvedaras V., Zemlys, V. \emph{Testing the functional constraints on parameters in regressions with variables of different frequency} Economics Letters 116 (2012) 250-254
}

