\name{gompertzp.gradient}
\alias{gompertzp.gradient}
\title{Gradient function for normalized Gompertz probability density function MIDAS weights specification
Calculate gradient function for normalized Gompertz probability density function specification of MIDAS weights.}
\usage{
  gompertzp.gradient(p, d)
}
\arguments{
  \item{p}{parameters for normalized Gompertz probability
  density function}

  \item{d}{number of coefficients}
}
\value{
  vector of coefficients
}
\description{
  Gradient function for normalized Gompertz probability
  density function MIDAS weights specification Calculate
  gradient function for normalized Gompertz probability
  density function specification of MIDAS weights.
}
\author{
  Julius Vainora
}

