% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{gompertzp_gradient}
\alias{gompertzp_gradient}
\title{Gradient function for normalized Gompertz probability density function MIDAS weights specification}
\usage{
gompertzp_gradient(p, d, m)
}
\arguments{
\item{p}{parameters for normalized Gompertz probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Calculate gradient function for normalized Gompertz probability density function specification of MIDAS weights.
}
\author{
Julius Vainora
}

