\name{lcauchyp.gradient}
\alias{lcauchyp.gradient}
\title{Gradient function for normalized log-Cauchy probability density function MIDAS weights specification
Calculate gradient function for normalized log-Cauchy probability density function specification of MIDAS weights.}
\usage{
  lcauchyp.gradient(p, d)
}
\arguments{
  \item{p}{parameters for normalized log-Cauchy probability
  density function}

  \item{d}{number of coefficients}
}
\value{
  vector of coefficients
}
\description{
  Gradient function for normalized log-Cauchy probability
  density function MIDAS weights specification Calculate
  gradient function for normalized log-Cauchy probability
  density function specification of MIDAS weights.
}
\author{
  Julius Vainora
}

