\name{lcauchyp}
\alias{lcauchyp}
\title{Normalized log-Cauchy probability density function MIDAS weights specification
Calculate MIDAS weights according to normalized log-Cauchy probability density function specification}
\usage{
  lcauchyp(p, d)
}
\arguments{
  \item{p}{parameters for normalized log-Cauchy probability
  density function}

  \item{d}{number of coefficients}
}
\value{
  vector of coefficients
}
\description{
  Normalized log-Cauchy probability density function MIDAS
  weights specification Calculate MIDAS weights according
  to normalized log-Cauchy probability density function
  specification
}
\author{
  Julius Vainora
}

