% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lcauchyp_gradient}
\alias{lcauchyp_gradient}
\title{Gradient function for normalized log-Cauchy probability density function MIDAS weights specification
Calculate gradient function for normalized log-Cauchy probability density function specification of MIDAS weights.}
\usage{
lcauchyp_gradient(p, d, m)
}
\arguments{
\item{p}{parameters for normalized log-Cauchy probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Gradient function for normalized log-Cauchy probability density function MIDAS weights specification
Calculate gradient function for normalized log-Cauchy probability density function specification of MIDAS weights.
}
\author{
Julius Vainora
}

