% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_nlpr.R
\name{lstr}
\alias{lstr}
\title{Compute LSTR term for high frequency variable}
\usage{
lstr(X, theta, beta, sd_x = sd(c(X), na.rm = TRUE))
}
\arguments{
\item{X}{matrix, high frequency variable embedded in low frequency, output of mls}

\item{theta}{vector, restriction coefficients for high frequency variable}

\item{beta}{vector of length 4, parameters for LSTR term, slope and 3 LSTR parameters}

\item{sd_x}{vector of length 1, defaults to standard deviation of X.}
}
\value{
a vector
}
\description{
Compute LSTR term for high frequency variable
}
