\name{mdslag}
\alias{mdslag}
\title{Lag a mixed data sampling time series}
\usage{
  mdslag(x, ...)
}
\arguments{
  \item{x}{a vector}

  \item{...}{further arguments}
}
\value{
  a matrix containing the lags
}
\description{
  Compute a lagged version of mixed data sampling time
  series
}
\author{
  Vaidotas Zemlys
}
\seealso{
  mdslag.default mdslag.mds
}

