\name{mdslag_coef}
\alias{mdslag_coef}
\title{Return the coefficients for mdslag variables}
\usage{
  mdslag_coef(x)
}
\arguments{
  \item{x}{an output from \code{\link{midas_u}}}
}
\value{
  a vector
}
\description{
  Extracts the coefficients and returns those coefficients
  which name has string \code{mdslag} in it.
}
\author{
  Vaidotas Zemlys
}

