\name{midas_coef}
\alias{midas_coef}
\title{Return the coefficients of MIDAS regression}
\usage{
  midas_coef(x)
}
\arguments{
  \item{x}{an output from \code{\link{midas_r}}}
}
\value{
  vector with coefficients of MIDAS regression
}
\description{
  A helper function for working with output of
  \code{\link{midas_r}}. Returns the regression
  coefficients.
}
\author{
  Vaidotas Zemlys
}

