% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_nlpr.R
\name{midas_lstr_simple}
\alias{midas_lstr_simple}
\title{LSTR  MIDAS regression}
\usage{
midas_lstr_simple(y, X, z = NULL, weight, start_lstr, start_x,
  start_z = NULL, method = c("Nelder-Mead", "BFGS"), ...)
}
\arguments{
\item{y}{model response}

\item{X}{prepared matrix of high frequency variable lags for LSTR term}

\item{z}{additional low frequency variables}

\item{weight}{the weight function}

\item{start_lstr}{the starting values for lstr term}

\item{start_x}{the starting values for weight function}

\item{start_z}{the starting values for additional low frequency variables}

\item{method}{a method passed to \link{optimx}}

\item{...}{additional parameters to \link{optimx}}
}
\value{
an object similar to \code{midas_r} object
}
\description{
Function for fitting LSTR MIDAS regression without the formula interface
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
