% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_nlpr.R
\name{midas_mmm_plain}
\alias{midas_mmm_plain}
\title{MMM (Mean-Min-Max)  MIDAS regression}
\usage{
midas_mmm_plain(y, X, z = NULL, weight, start_mmm, start_x, start_z = NULL,
  method = c("Nelder-Mead", "BFGS"), ...)
}
\arguments{
\item{y}{model response}

\item{X}{prepared matrix of high frequency variable lags for MMM term}

\item{z}{additional low frequency variables}

\item{weight}{the weight function}

\item{start_mmm}{the starting values for MMM term}

\item{start_x}{the starting values for weight function}

\item{start_z}{the starting values for additional low frequency variables}

\item{method}{a method passed to \link{optimx}}

\item{...}{additional parameters to \link{optimx}}
}
\value{
an object similar to \code{midas_r} object
}
\description{
Function for fitting MMM MIDAS regression without the formula interface
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
