% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_nlpr.R
\name{midas_nlpr.fit}
\alias{midas_nlpr.fit}
\title{Fit restricted MIDAS regression}
\usage{
midas_nlpr.fit(x)
}
\arguments{
\item{x}{\code{midas_r} object}
}
\value{
\code{\link{midas_r}} object
}
\description{
Workhorse function for fitting restricted MIDAS regression
}
\author{
Vaidotas Zemlys
}
