% Generated by roxygen2 (4.0.1): do not edit by hand
\name{midas_r.fit}
\alias{midas_r.fit}
\title{Fit restricted MIDAS regression}
\usage{
\method{midas_r}{fit}(x)
}
\arguments{
\item{x}{\code{midas_r} object}
}
\value{
\code{\link{midas_r}} object
}
\description{
Workhorse function for fitting restricted MIDAS regression
}
\author{
Vaidotas Zemlys
}

