\name{midas_r_fast}
\alias{midas_r_fast}
\title{Restricted MIDAS regression}
\usage{
  midas_r_fast(y, X, z = NULL, weight, grw = NULL, startx,
    startz = NULL, method = "BFGS", ...)
}
\arguments{
  \item{y}{}

  \item{X}{}

  \item{z}{}

  \item{weight}{}

  \item{grw}{}

  \item{startx}{}

  \item{startz}{}

  \item{method}{}

  \item{...}{}
}
\value{
  a \code{midas_r} object
}
\description{
  Fast function for fitting MIDAS regression
}
\author{
  Virmantas Kvedars, Vaidotas Zemlys
}

