\name{mls_coef}
\alias{mls_coef}
\title{Return the coefficients for fmls variables}
\usage{
mls_coef(x)
}
\arguments{
  \item{x}{an output from \code{\link{midas_u}}}
}
\value{
a vector
}
\description{
Extracts the coefficients and returns those coefficients
which name has string \code{fmls} or \code{mls} in it.
}
\author{
Vaidotas Zemlys
}

