% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_nlpr.R
\name{mmm}
\alias{mmm}
\title{Compute MMM term for high frequency variable}
\usage{
mmm(X, theta, beta)
}
\arguments{
\item{X}{matrix, high frequency variable embedded in low frequency, output of mls}

\item{theta}{vector, restriction coefficients for high frequency variable}

\item{beta}{vector of length 2, parameters for MMM term, slope and MMM parameter.}
}
\value{
a vector
}
\description{
Compute MMM term for high frequency variable
}
