\name{nakagamip}
\alias{nakagamip}
\title{Normalized Nakagami probability density function MIDAS weights specification
Calculate MIDAS weights according to normalized Nakagami probability density function specification}
\usage{
  nakagamip(p, d)
}
\arguments{
  \item{p}{parameters for normalized Nakagami probability
  density function}

  \item{d}{number of coefficients}
}
\value{
  vector of coefficients
}
\description{
  Normalized Nakagami probability density function MIDAS
  weights specification Calculate MIDAS weights according
  to normalized Nakagami probability density function
  specification
}
\author{
  Julius Vainora
}

