% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nakagamip_gradient}
\alias{nakagamip_gradient}
\title{Gradient function for normalized Nakagami probability density function MIDAS weights specification
Calculate gradient function for normalized Nakagami probability density function specification of MIDAS weights.}
\usage{
nakagamip_gradient(p, d, m)
}
\arguments{
\item{p}{parameters for normalized Nakagami probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Gradient function for normalized Nakagami probability density function MIDAS weights specification
Calculate gradient function for normalized Nakagami probability density function specification of MIDAS weights.
}
\author{
Julius Vainora
}

