% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nbetaMT.gradient}
\alias{nbetaMT.gradient}
\title{Gradient function for normalized beta probability density function MIDAS weights specification (MATLAB toolbox compatible)
Calculate gradient function for normalized beta probability density function specification of MIDAS weights.}
\usage{
nbetaMT.gradient(p, d, m)
}
\arguments{
\item{p}{parameters for normalized beta probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Gradient function for normalized beta probability density function MIDAS weights specification (MATLAB toolbox compatible)
Calculate gradient function for normalized beta probability density function specification of MIDAS weights.
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}

