% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{nbetaMT_gradient}
\alias{nbetaMT_gradient}
\title{Gradient function for normalized beta probability density function MIDAS weights specification (MATLAB toolbox compatible)}
\usage{
nbetaMT_gradient(p, d, m)
}
\arguments{
\item{p}{parameters for normalized beta probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Calculate gradient function for normalized beta probability density function specification of MIDAS weights.
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
