\name{nealmon.gradient}
\alias{nealmon.gradient}
\title{Gradient function for normalized exponential Almon lag weights}
\usage{
nealmon.gradient(p, d, m)
}
\arguments{
  \item{p}{hyperparameters for Almon lag}

  \item{d}{number of coefficients}

  \item{m}{the frequency ratio, currently ignored}
}
\value{
the gradient matrix
}
\description{
Gradient function for normalized exponential Almon lag
weights
}
\author{
Vaidotas Zemlys
}

