% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lagspec.R
\name{nealmon_gradient}
\alias{nealmon_gradient}
\title{Gradient function for normalized exponential Almon lag weights}
\usage{
nealmon_gradient(p, d, m)
}
\arguments{
\item{p}{hyperparameters for Almon lag}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
the gradient matrix
}
\description{
Gradient function for normalized exponential Almon lag weights
}
\author{
Vaidotas Zemlys
}

