% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_nlpr_methods.R
\name{plot_lstr}
\alias{plot_lstr}
\title{Plot MIDAS coefficients}
\usage{
plot_lstr(x, term_name, title = NULL, compare = NULL, ...)
}
\arguments{
\item{x}{\code{midas_r} object}

\item{term_name}{the term name for which the coefficients are plotted. Default is \code{NULL}, which selects the first MIDAS term}

\item{title}{the title string of the graph. The default is \code{NULL} for the default title.}

\item{compare}{the parameters for weight function to compare with the model, default is NULL}

\item{...}{not used}
}
\value{
a data frame with restricted MIDAS coefficients, unrestricted MIDAS coefficients and lower and upper confidence interval limits. The data
frame is returned invisibly.
}
\description{
Plots logistic function for LSTR MIDAS regression
}
\details{
Plots logistic function for LSTR MIDSAS regression
of unrestricted MIDAS regression
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
