% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_sp_methods.R
\name{plot_sp}
\alias{plot_sp}
\title{Plot non-parametric part of the single index MIDAS regression}
\usage{
plot_sp(x, term_name, title = NULL, compare = NULL, ...)
}
\arguments{
\item{x}{\code{midas_r} object}

\item{term_name}{the term name for which the coefficients are plotted. Default is \code{NULL}, which selects the first MIDAS term}

\item{title}{the title string of the graph. The default is \code{NULL} for the default title.}

\item{compare}{the parameters for weight function to compare with the model, default is NULL}

\item{...}{not used}
}
\value{
a data frame with restricted MIDAS coefficients, unrestricted MIDAS coefficients and lower and upper confidence interval limits. The data
frame is returned invisibly.
}
\description{
Plot non-parametric part of the single index MIDAS regression
of unrestricted MIDAS regression
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
