% Generated by roxygen2 (4.0.1): do not edit by hand
\name{polystep.gradient}
\alias{polystep.gradient}
\title{Gradient of step function specification for MIDAS weights}
\usage{
polystep.gradient(p, d, m, a)
}
\arguments{
\item{p}{vector of parameters}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}

\item{a}{vector of increasing positive integers indicating the steps}
}
\value{
vector of coefficients
}
\description{
Gradient of step function specification for MIDAS weights
}
\author{
Vaidotas Zemlys
}

