\name{polystep}
\alias{polystep}
\title{Step function specification for MIDAS weights}
\usage{
  polystep(p, d, a)
}
\arguments{
  \item{p}{vector of parameters}

  \item{d}{number of coefficients}

  \item{a}{vector of increasing positive integers
  indicating the steps}
}
\value{
  vector of coefficients
}
\description{
  Step function specification for MIDAS weights
}
\author{
  Vaidotas Zemlys
}

