\name{residuals.midas_r}
\alias{residuals.midas_r}
\title{Calculate residuals of MIDAS regression}
\usage{
  residuals.midas_r(x)
}
\arguments{
  \item{x}{a \code{\link{midas_r}} object}
}
\value{
  a vector of residuals of MIDAS regression
}
\description{
  Calculate residuals of MIDAS regression
}
\author{
  Vaidotas Zemlys
}

