\name{residuals.midas_r}
\alias{residuals.midas_r}
\title{Calculate residuals of MIDAS regression}
\usage{
  \method{residuals}{midas_r} (object, ...)
}
\arguments{
  \item{object}{a \code{\link{midas_r}} object}

  \item{...}{currently nothing}
}
\value{
  a vector of residuals of MIDAS regression
}
\description{
  Calculate residuals of MIDAS regression
}
\author{
  Vaidotas Zemlys
}

