\name{restr_coef}
\alias{restr_coef}
\title{Return the restricted coefficients generated by restriction function(s)}
\usage{
  restr_coef(x, name = restr_names(x))
}
\arguments{
  \item{x}{an output from \code{\link{midas_r}}}

  \item{name}{name(s) of the restriction function(s), the
  default value is the name(s) of the restriction
  function(s) supplied to \code{\link{midas_r}}}
}
\value{
  a list if \code{length(name)>1}, a vector otherwise
}
\description{
  A helper function for working with output of
  \code{\link{midas_r}}. Returns the restricted
  coefficients generated by restriction function(s)
}
\author{
  Vaidotas Zemlys
}

