% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{data}
\name{rvsp500}
\alias{rvsp500}
\title{Realized volatility of S&P500 index}
\format{
A \code{data.frame} object with two columns. First column contains date id, and the second the realized volatility for S&P500 index.
}
\source{
No longer available. Read the statement here: \href{https://oxford-man.ox.ac.uk/research/realized-library/}{https://oxford-man.ox.ac.uk/research/realized-library/}
}
\description{
Realized volatility of S&P500(Live) index of the period 2000 01 03 - 2013 11 22
}
\examples{
## Do not run:
## The original data contained the file the file OxfordManRealizedVolatilityIndices.csv. The code below 
## reproduces the dataset.

## rvi <- read.csv("OxfordManRealizedVolatilityIndices.csv",check.names=FALSE,skip=2)
## ii <- which(rvi$DateID=="20131112")
## rvsp500 <- na.omit(rvi[1:ii,c("DataID","SPX2.rv")]
}
\references{
Heber, Gerd and Lunde, Asger, and Shephard, Neil and Sheppard, Kevin \emph{Oxford-Man Institute's realized library}, Oxford-Man Institute, University of Oxford (2009)
}
\keyword{datasets}
