\name{summary.midas_r}
\alias{summary.midas_r}
\title{Summarizing MIDAS regression fit}
\usage{
  \method{summary}{midas_r} (x)
}
\arguments{
  \item{x}{\code{\link{midas_r}} object}
}
\value{
  The function \code{summary.midas_r} computes a list of
  summary statistics of fitted model given in \code{x}
}
\description{
  summary method for class \code{midas_r}
}
\author{
  Virmantas Kvedaras, Vaidotas Zemlys
}

