% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/midasreg.R
\name{update_weights}
\alias{update_weights}
\title{Updates weights in MIDAS regression formula}
\usage{
update_weights(expr, tb)
}
\arguments{
\item{expr}{expression with MIDAS term}

\item{tb}{a named list with redefined weights}
}
\value{
an expression with changed weights
}
\description{
Updates weights in a expression with MIDAS term
}
\details{
For a MIDAS term \code{fmls(x, 6, 1, nealmon)} change weight \code{nealmon} to another weight.
}
\examples{
update_weights(y~trend+mls(x,0:7,4,nealmon)+mls(z,0:16,12,nealmon),list(x = "nbeta", z = ""))
}
\author{
Vaidotas Zemlys
}

