% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{data}
\name{UScpiqs}
\alias{UScpiqs}
\title{US quartely seasonaly adjusted consumer price index}
\format{
A \code{\link{data.frame}} object.
}
\source{
\href{http://www.bea.gov/national/xls/gdplev.xls}{FRED}
}
\description{
US quarterly CPI from 1960Q1 to 2017Q3s
}
\keyword{datasets}
