% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{data}
\name{USeffrw}
\alias{USeffrw}
\title{US weekly effective federal funds rate.}
\format{
A \code{\link{data.frame}} object.
}
\source{
\href{http://www.bea.gov/national/xls/gdplev.xls}{FRED}
}
\description{
US weekly effective federal funds rate from 1954-07-07 to 2017-12-13
}
\keyword{datasets}
