% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{data}
\name{USpayems}
\alias{USpayems}
\title{United States total employment non-farms payroll, monthly, seasonally adjusted.}
\format{A \code{\link{ts}} object.}
\source{
\href{http://research.stlouisfed.org/fred2/series/PAYEMS}{FRED, Federal Reserve Economic Data, from the Federal Reserve Bank of St. Louis}
}
\description{
United States total employment non-farms payroll, monthly, seasonally adjusted. Retrieved from FRED, symbol "PAYEMS" at 2014-04-25.
}
\examples{

## Do not run:
## library(quantmod)
## USpayems <- ts(getSymbols("PAYEMS",src="FRED",auto.assign=FALSE),start=c(1939,1),frequency=12)
}
\keyword{datasets}

