% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{data}
\name{USqgdp}
\alias{USqgdp}
\title{United States gross domestic product, quarterly, seasonaly adjusted annual rate.}
\format{A \code{\link{ts}} object.}
\source{
\href{http://research.stlouisfed.org/fred2/series/GDP}{FRED, Federal Reserve Economic Data, from the Federal Reserve Bank of St. Louis}
}
\description{
United States gross domestic product, quarterly, seasonaly adjusted annual rate. Retrieved from FRED, symbol "GDP" at 2014-04-25.
}
\examples{
## Do not run:
## library(quantmod)
## USqgdp <- ts(getSymbols("GDP",src="FRED",auto.assign=FALSE),start=c(1947,1),frequency=4)
}
\keyword{datasets}

