% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{data}
\name{USunempr}
\alias{USunempr}
\title{US monthly unemployment rate}
\format{
A \code{\link{ts}} object.
}
\source{
\href{https://fred.stlouisfed.org/series/UNRATE}{FRED}
}
\description{
The monthly unemployment rate for United States from 1948 to 2011.
}
\keyword{datasets}
