% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weight_names}
\alias{weight_names}
\title{Return the names of restriction function(s)}
\usage{
weight_names(x)
}
\arguments{
\item{x}{an output from \code{\link{midas_r}}}
}
\value{
a character vector
}
\description{
A helper function for working with output of \code{\link{midas_r}}. Returns the name(s) of restriction function(s) used in call to \code{\link{midas_r}}
}
\author{
Vaidotas Zemlys
}

