% Generated by roxygen2 (4.0.1): do not edit by hand
\name{almonp.gradient}
\alias{almonp.gradient}
\title{Gradient function for Almon polynomial MIDAS weights}
\usage{
almonp.gradient(p, d, m)
}
\arguments{
\item{p}{vector of parameters for Almon polynomial specification}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Calculate gradient for Almon polynomial MIDAS weights specification
}
\author{
Vaidotas Zemlys
}

