% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amweights}
\alias{amweights}
\title{Weights for aggregates based MIDAS regressions}
\usage{
amweights(p, d, m, weight = nealmon, type = c("A", "B", "C"))
}
\arguments{
\item{p}{parameters for weight functions, see details.}

\item{d}{number of high frequency lags}

\item{m}{the frequency}

\item{weight}{the weight function}

\item{type}{type of structure, a string, one of A, B or C.}
}
\value{
a vector of weights
}
\description{
Produces weights for aggregates based MIDAS regression
}
\details{
Suppose a weight function \eqn{w(\beta,\theta)} satisfies the following equation:
\deqn{w(\beta,\theta)=\beta g(\theta)}

The following combinations are defined, corresponding to structure types \code{A}, \code{B} and \code{C} respectively:
\deqn{(w(\beta_1,\theta_1),...,w(\beta_k,\theta_k))}
\deqn{(w(\beta_1,\theta),...,w(\beta_k,\theta))}
\deqn{\beta(w(1,\theta_1),...,w(1,\theta_k))}

The starting values \eqn{p} should be supplied then as follows:
\deqn{(\beta_1,\theta_1,...,\beta_k,\theta_k)}
\deqn{(\beta_1,...,\beta_k,\theta)}
\deqn{(\beta,\theta_1,...,\theta_k)}
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}

