% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/midaslag.R
\name{check_mixfreq}
\alias{check_mixfreq}
\title{Check data for MIDAS regression}
\usage{
check_mixfreq(data)
}
\arguments{
\item{data}{a list containing mixed frequency data}
}
\value{
a boolean TRUE, if mixed frequency data is conformable, FALSE if it is not.
}
\description{
Given mixed frequency data check whether higher frequency data can be converted to the lowest frequency.
}
\details{
The number of observations in higher frequency data elements should have a common divisor with the number of observations in response variable. It is always assumed that the response variable is of the lowest frequency.
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}

