\name{checkARstar}
\alias{checkARstar}
\title{Check whether the MIDAS model is MIDAS-AR* model}
\usage{
  checkARstar(x, env, data)
}
\arguments{
  \item{x}{the model formula}

  \item{env}{the environment of model}

  \item{data}{the model data}
}
\description{
  Checks whether the MIDAS model is MIDAS-AR* model and
  provides necessary modifications
}
\author{
  Julius Vainora
}

