\name{checkARstar}
\alias{checkARstar}
\title{Check whether the MIDAS model is MIDAS-AR* model}
\usage{
  checkARstar(trms)
}
\arguments{
  \item{trms}{terms of the model formula}
}
\description{
  Checks whether the MIDAS model is MIDAS-AR* model and
  provides necessary modifications
}
\author{
  Julius Vainora
}

