\name{forecast.midas_r}
\alias{forecast.midas_r}
\title{Forecast MIDAS regression}
\usage{
  \method{forecast}{midas_r} (object, newdata)
}
\arguments{
  \item{x}{midas_r object}

  \item{newdata}{newdata}
}
\value{
  a vector of forecasts
}
\description{
  Forecasts MIDAS regression. Differs from \code{predict},
  that it respects history
}
\details{
  Add later
}
\author{
  Vaidotas Zemlys
}

