\name{gompertzp}
\alias{gompertzp}
\title{Normalized Gompertz probability density function MIDAS weights specification
Calculate MIDAS weights according to normalized Gompertz probability density function specification}
\usage{
  gompertzp(p, d)
}
\arguments{
  \item{p}{parameters for normalized Gompertz probability
  density function}

  \item{d}{number of coefficients}
}
\value{
  vector of coefficients
}
\description{
  Normalized Gompertz probability density function MIDAS
  weights specification Calculate MIDAS weights according
  to normalized Gompertz probability density function
  specification
}
\author{
  Julius Vainora
}

