% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{gompertzp}
\alias{gompertzp}
\title{Normalized Gompertz probability density function MIDAS weights specification}
\usage{
gompertzp(p, d, m)
}
\arguments{
\item{p}{parameters for normalized Gompertz probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Calculate MIDAS weights according to normalized Gompertz probability density function specification
}
\author{
Julius Vainora
}

