% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{lcauchyp}
\alias{lcauchyp}
\title{Normalized log-Cauchy probability density function MIDAS weights specification
Calculate MIDAS weights according to normalized log-Cauchy probability density function specification}
\usage{
lcauchyp(p, d, m)
}
\arguments{
\item{p}{parameters for normalized log-Cauchy probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Normalized log-Cauchy probability density function MIDAS weights specification
Calculate MIDAS weights according to normalized log-Cauchy probability density function specification
}
\author{
Julius Vainora
}

