\name{mdslag}
\alias{mdslag}
\alias{mdslag.default}
\alias{mdslag.mds}
\title{Lag a mixed data sampling time series}
\usage{
  mdslag(x, ...)
}
\arguments{
  \item{x}{a vector}

  \item{...}{}
}
\value{
  a matrix containing the lags
}
\description{
  Compute a lagged version of mixed data sampling time
  series
}
\author{
  Vaidotas Zemlys
}

