% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{package}
\name{midasr-package}
\alias{midasr}
\alias{midasr-package}
\title{Mixed Data Sampling Regression}
\description{
Package for estimating, testing and forecasting MIDAS regression.
}
\details{
Methods and tools for mixed frequency time series data analysis. Allows estimation, model selection and forecasting for MIDAS regressions.
}
\author{
Virmantas Kvedaras \email{virmantas.kvedaras@mif.vu.lt}, Vaidotas Zemlys (maintainer) \email{zemlys@gmail.com}
}
\keyword{package}

