% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{midasr-package}
\alias{midasr}
\alias{midasr-package}
\title{Estimating and testing MIDAS regression}
\description{
Package for estimating and testing MIDAS regression.
}
\details{
The main feature of this package is function \code{\link{hAh.test}}
which performs test whether coefficients of MIDAS regression have certain functional form
}
\author{
Virmantas Kvedaras \email{virmantas.kvedaras@mif.vu.lt}, Vaidotas Zemlys (maintainer) \email{zemlys@gmail.com}
}
\keyword{package}

