% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midaslag.R
\name{mlsd}
\alias{mlsd}
\title{MIDAS lag structure with dates}
\usage{
mlsd(x, k, datex, datey, ...)
}
\arguments{
\item{x}{a vector}

\item{k}{lags, a vector}

\item{datex}{a vector of high frequency dates, if}

\item{datey}{low frequency dates}

\item{...}{further arguments used in fitting MIDAS regression}
}
\value{
a matrix containing the first differences and the lag k+1.
}
\description{
MIDAS lag structure with dates
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys-Balevičius
}
