% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{nakagamip}
\alias{nakagamip}
\title{Normalized Nakagami probability density function MIDAS weights specification}
\usage{
nakagamip(p, d, m)
}
\arguments{
\item{p}{parameters for normalized Nakagami probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Calculate MIDAS weights according to normalized Nakagami probability density function specification
}
\author{
Julius Vainora
}

