\name{nbeta.gradient}
\alias{nbeta.gradient}
\title{Gradient function for normalized beta probability density function MIDAS weights specification
Calculate gradient function for normalized beta probability density function specification of MIDAS weights.}
\usage{
  nbeta.gradient(p, d)
}
\arguments{
  \item{p}{parameters for normalized beta probability
  density function}

  \item{d}{number of coefficients}
}
\value{
  vector of coefficients
}
\description{
  Gradient function for normalized beta probability density
  function MIDAS weights specification Calculate gradient
  function for normalized beta probability density function
  specification of MIDAS weights.
}
\author{
  Virmantas Kvedaras, Vaidotas Zemlys
}

