% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagspec.R
\name{nbeta}
\alias{nbeta}
\title{Normalized beta probability density function MIDAS weights specification
Calculate MIDAS weights according to normalized beta probability density function specification}
\usage{
nbeta(p, d, m)
}
\arguments{
\item{p}{parameters for normalized beta probability density function}

\item{d}{number of coefficients}

\item{m}{the frequency ratio, currently ignored}
}
\value{
vector of coefficients
}
\description{
Normalized beta probability density function MIDAS weights specification
Calculate MIDAS weights according to normalized beta probability density function specification
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
