\name{nealmon.gradient}
\alias{nealmon.gradient}
\title{Gradient function for normalized exponential Almon lag weights}
\usage{
  nealmon.gradient(p, d)
}
\arguments{
  \item{p}{hyperparameters for Almon lag}

  \item{d}{number of coefficients}
}
\value{
  the gradient matrix
}
\description{
  Gradient function for normalized exponential Almon lag
  weights
}
\author{
  Vaidotas Zemlys
}

