\name{predict.midas_r}
\alias{predict.midas_r}
\title{Predict method for MIDAS regression fit}
\usage{
  \method{predict}{midas_r} (object, newdata, ...)
}
\arguments{
  \item{object}{\code{\link{midas_r}} object}

  \item{newdata}{a named list containing data for mixed
  frequencies}

  \item{...}{additional arguments, not used}
}
\value{
  a vector of predicted values
}
\description{
  Predicted values based on MIDAS regression object
}
\author{
  Virmantas Kvedaras, Vaidotas Zemlys
}

