% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{prep_hAh}
\alias{prep_hAh}
\title{Calculate data for \link{hAh_test} and \link{hAhr_test}}
\usage{
prep_hAh(x)
}
\arguments{
\item{x}{\code{midas_r} object}
}
\value{
a list with necessary matrices
}
\description{
Workhorse function for calculating necessary matrices for \link{hAh_test} and \link{hAhr_test}. Takes the same parameters as \link{hAh_test}
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}
\seealso{
hAh_test, hAhr_test
}

