\docType{data}
\name{rvsp500}
\alias{rvsp500}
\title{Realized volatility of S&P500 index}
\format{A \code{data.frame} object with two columns. First column contains date id, and the second the realized volatility for S&P500 index.}
\source{
\href{http://realized.oxford-man.ox.ac.uk/media/1366/oxfordmanrealizedvolatilityindices.zip}{Heber,
Gerd, Asger Lunde, Neil Shephard and Kevin Sheppard (2009)
"Oxford-Man Institute's realized library", Oxford-Man
Institute, University of Oxfors}
}
\description{
Realized volatility of S&P500(Live) index of the period
2000 01 03 - 2012 05 22
}
\keyword{datasets}

