% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{data}
\name{UScpiqs}
\alias{UScpiqs}
\title{US quartely seasonaly adjusted consumer price index}
\format{
A \code{\link{data.frame}} object.
}
\source{
\href{https://fred.stlouisfed.org/series/CPALTT01USQ661S}{FRED}
}
\description{
US quarterly CPI from 1960Q1 to 2017Q3s. Seasonaly adjusted, Index 2015=1
}
\keyword{datasets}
