% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{USrealgdp}
\alias{USrealgdp}
\title{US annual gross domestic product in billions of chained 2005 dollars}
\format{A \code{\link{ts}} object.}
\source{
\href{http://www.bea.gov/national/xls/gdplev.xls}{U.S. Department of Commerce, Bureau of Economic Analysis}
}
\description{
The annual gross domestic product in billions of chained 2005 dollars for US from 1948 to 2011.
}
\keyword{datasets}

