% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasr-package.R
\docType{data}
\name{USunempr}
\alias{USunempr}
\title{US monthly unemployment rate}
\format{
A \code{\link{ts}} object.
}
\source{
\href{http://data.bls.gov/timeseries/LNS14000000}{U.S. Bureau of Labor Statistics}
}
\description{
The monthly unemployment rate for United States from 1948 to 2011.
}
\keyword{datasets}
